+++ yellowstone-grpc-proto/src/plugin/filter/mod.rs
@@ -1,6 +1,7 @@
 #[allow(clippy::module_inception)]
-pub mod filter;
+mod filter;
 pub mod limits;
+pub mod message;
 pub mod name;
 
-pub use {filter::*, limits::*, name::*};
+pub use filter::{Filter, FilterAccountsDataSlice, FilterError, FilterResult};
